Refactor out functions
