+++ app/utils/product.ts
@@ -1,3 +1,4 @@
+import invariant from "tiny-invariant";
 import type {
   FullProduct,
   SelectedProductOption,
@@ -55,4 +56,68 @@ export function getProductOptionIcon(
     );
 
   return matchingOption?.icon;
+}
+
+export function getSelectedAccessories(
+  fullMatchingAccessories: (FullProduct | undefined)[],
+  serializedFormDataQuantities: {
+    name: string;
+    value: FormDataEntryValue;
+  }[]
+): {
+  variantId: string;
+  quantity: number;
+}[] {
+  return fullMatchingAccessories.map((matchingAccessory) => {
+    invariant(matchingAccessory, "matchingAccessory not found");
+    const matchingQuantity = serializedFormDataQuantities.find(
+      (quantity) => quantity.name === matchingAccessory.id
+    );
+
+    return {
+      variantId: matchingAccessory.defaultVariantId,
+      quantity: matchingQuantity
+        ? parseInt(matchingQuantity.value.toString())
+        : 0,
+    };
+  });
+}
+
+export function getSelectedAccessoriesWithOptions(
+  fullMatchingAccessories: (FullProduct | undefined)[],
+  serializedFormDataOptionQuantities: {
+    name: string;
+    value: string;
+    quantity: number;
+    accessoryId: string;
+  }[]
+): {
+  variantId: string;
+  quantity: number;
+}[] {
+  return fullMatchingAccessories.flatMap((matchingAccessory) => {
+    invariant(matchingAccessory, "matchingAccessory not found");
+
+    const matchingOptions = serializedFormDataOptionQuantities.filter(
+      (optionQuantity) => optionQuantity.accessoryId === matchingAccessory.id
+    );
+
+    return matchingOptions.map((option) => {
+      const matchingQuantity = serializedFormDataOptionQuantities.find(
+        (quantity) => quantity.value === option.value
+      );
+
+      let matchingVariant = getMatchingVariant(
+        [{ name: option.name, value: option.value }],
+        matchingAccessory
+      );
+
+      return {
+        quantity: matchingQuantity ? matchingQuantity.quantity : 0,
+        variantId: matchingVariant
+          ? matchingVariant.id
+          : matchingAccessory.defaultVariantId,
+      };
+    });
+  });
 }
